Various improvements to avoid NPE and refactoring
